Adicionando retorno 404

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -1,8 +1,8 @@
 package br.com.alura.forum.controller;
 
-
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -27,62 +27,71 @@ import br.com.alura.forum.modelo.Topico;
 import br.com.alura.forum.repository.CursoRepository;
 import br.com.alura.forum.repository.TopicoRepository;
 
-
 @RestController
 @RequestMapping("/topicos")
 public class TopicosController {
-	
+
 	@Autowired
 	private TopicoRepository topicosRepository;
-	
+
 	@Autowired
 	private CursoRepository cursoRepository;
-	
+
 	// Cada metodo roda dentro de uma transação
 	@GetMapping
-	public List<TopicoDto> lista(String nomeCurso){
-		if(nomeCurso == null) {
+	public List<TopicoDto> lista(String nomeCurso) {
+		if (nomeCurso == null) {
 			List<Topico> topicos = topicosRepository.findAll();
 			return TopicoDto.converter(topicos);
-		}else {
+		} else {
 			List<Topico> topicos = topicosRepository.findByCurso_Nome(nomeCurso);
 			return TopicoDto.converter(topicos);
 		}
 	}
-	
+
 	@PostMapping
 	@Transactional
 	public ResponseEntity<TopicoDto> cadastrar(@RequestBody @Valid TopicoForm form, UriComponentsBuilder uriBuilder) {
 		Topico topico = form.converter(cursoRepository);
 		topicosRepository.save(topico);
-		
+
 		URI uri = uriBuilder.path("/topicos/{id}").buildAndExpand(topico.getId()).toUri();
-		
+
 		return ResponseEntity.created(uri).body(new TopicoDto(topico));
 	}
-	
+
 	@GetMapping("/{id}")
-	public DetalhesDoTopicoDto detalhar(@PathVariable("id") Long codigo) {
-		Topico one = topicosRepository.getOne(codigo);
-		return new DetalhesDoTopicoDto(one);
+	public ResponseEntity<?> detalhar(@PathVariable("id") Long codigo) {
+		Optional<Topico> one = topicosRepository.findById(codigo);
+		if (one.isPresent()) {
+			return ResponseEntity.ok(new DetalhesDoTopicoDto(one.get()));
+		}
+		return ResponseEntity.notFound().build();
 	}
-		
+
 	@PutMapping("/{id}")
-	// Não há necessidade de chamar o repository denovo, pois como a classe topico se tornou managed
+	// Não há necessidade de chamar o repository denovo, pois como a classe topico
+	// se tornou managed
 	// A atualização no banco de dados será realizada automaticamente.
 	@Transactional
-	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody AtualizacaoTopicoForm form){
-		
-		
-		Topico topico = form.atualizar(id, topicosRepository);
-		return ResponseEntity.ok(new TopicoDto(topico));
+	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody AtualizacaoTopicoForm form) {
+		Optional<Topico> optional = topicosRepository.findById(id);
+		if (optional.isPresent()) {
+			Topico topico = form.atualizar(id, topicosRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
+
+		return ResponseEntity.notFound().build();
 	}
-	
+
 	@DeleteMapping("{id}")
 	@Transactional
-	public ResponseEntity<?> remover(@PathVariable Long id){
-		topicosRepository.deleteById(id);
-		return ResponseEntity.ok().build();
+	public ResponseEntity<?> remover(@PathVariable Long id) {
+		Optional<Topico> optional = topicosRepository.findById(id);
+		if (optional.isPresent()) {
+			topicosRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+		return ResponseEntity.notFound().build();
 	}
-	
 }
